- It compiles! - Fix of a invalid memory access on Mpi. - Removed the branch in bn_mod1_low. - Refactor on the way bn_coefs is used. Now we alloc previously d_bn_coefs and h_bn_coefs (this one is just an auxiliar array). Doing this we get speed on update_*_data(). - Refactor on NTT. Prime P replaced by 4293918721.

+++ src/cuda_functions.h
@@ -79,9 +79,9 @@ class CUDAFunctions{
   private:
 };
 #ifdef NTTMUL
-__device__ __host__ inline uint64_t s_rem (uint64_t a);
-__device__ __host__  uint64_t s_mul(uint64_t a,
-                                    uint64_t b);
+__device__ __host__ inline uint32_t s_rem (uint64_t a);
+__device__ __host__  uint32_t s_mul(uint32_t a,
+                                    uint32_t b);
 #endif
 
 #endif
